+++ bottle/src/database/user.rs
@@ -1,7 +1,8 @@
-use bottle_orm::{FromAnyRow, Model};
+use bottle_orm::Model;
 use chrono::{DateTime, Utc};
+use serde::{Deserialize, Serialize};
 
-#[derive(Debug, Model, FromAnyRow)]
+#[derive(Debug, Model, Serialize, Deserialize)]
 pub struct User {
     #[orm(primary_key)]
     pub id: String,
@@ -13,13 +14,12 @@ pub struct User {
     pub updated_at: DateTime<Utc>,
 }
 
-#[derive(Debug, Model, FromAnyRow)]
+#[derive(Debug, Model, Serialize, Deserialize)]
 pub struct Account {
-	#[orm(primary_key)]
-	pub id: String,
-	#[orm(foreign_key = "User::id")]
-	pub user_id: String,
-	pub account_type: String,
+        #[orm(primary_key)]
+        pub id: String,
+        #[orm(foreign_key = "User::id")]
+        pub user_id: String,	pub account_type: String,
 	pub password: String,
 	pub changed_password: DateTime<Utc>,
 	pub created_at: DateTime<Utc>